+++ src/app/budget/models/budget.model.ts
@@ -1,8 +1,10 @@
+import { Timestamp } from '@firebase/firestore-types';
+
 export interface Budget {
     id: string;
     name: string;
-    dateStart: Date;
-    dateEnd: Date;
+    dateStart: Timestamp;
+    dateEnd: Timestamp;
     totalCash: number;
     cashLeft: number;
 }