+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.35.2
+* Animations on routing - improved tabs styling
+* Hammerjs for BudgetMain component - fixed scrolling issue
+
 ## 0.35.0
 * Animations on routing
 * Hammerjs for BudgetMain component